==> 2. pivot-notused.ipynb <==
import pandas as pd
import os

# IMPORT ALL .CSV
provinces = ['Bangkok','Chanthaburi','Chiang Mai','Kanchanaburi','Khon Kaen','Songkhla']
data = dict()
subset = 'Train'

for province in provinces:
    df =  pd.read_csv(f'./full-data/{subset}/{province}_full.csv', index_col=0, parse_dates=True)
    idx = pd.MultiIndex.from_product([[province],df.columns])
    df.columns = idx
    data[province] = df



grand_data = data['Bangkok']\
                .merge(data['Chanthaburi'], how='left', left_index=True, right_index=True)\
                .merge(data['Chiang Mai'], how='left', left_index=True, right_index=True)\
                .merge(data['Kanchanaburi'], how='left', left_index=True, right_index=True)\
                .merge(data['Khon Kaen'], how='left', left_index=True, right_index=True)\
                .merge(data['Songkhla'], how='left', left_index=True, right_index=True)
grand_data.head()

grand_data = data['Bangkok']\
                .merge(data['Chanthaburi'], how='left', left_index=True, right_index=True)\
                .merge(data['Chiang Mai'], how='left', left_index=True, right_index=True)\
                .merge(data['Kanchanaburi'], how='left', left_index=True, right_index=True)\
                .merge(data['Khon Kaen'], how='left', left_index=True, right_index=True)\
                .merge(data['Songkhla'], how='left', left_index=True, right_index=True)
grand_data.head()

if not os.path.exists('pivot_.csv'):    
    grand_data.stack(0, dropna=False).to_csv('pivot_.csv')
    print('Create new file')
