==> 3. fillna.ipynb <==
from collections import defaultdict
from util_handle_missing import *
import pandas as pd
import os

provinces = ['Bangkok','Chanthaburi','Chiang Mai','Kanchanaburi','Khon Kaen','Songkhla']

data = defaultdict(lambda: {})

for subset in ['Train','Test']:
    for p in provinces:
        df = pd.read_csv(f'full-data/{subset}/{p}_full.csv', 
                            index_col=0, parse_dates=True)
        data[subset][p] = df
        print(f"{subset} : {p} : {df.shape}")

for subset in ['Train','Test']:
    print(subset)
    for p in provinces:
        print(f"\t{p} : \n{data[subset][p].isnull().sum()}")
    print("-"*30)

# ===========================================================
# Nan = zero
ตอนหลังเขาแก้กลับอีกแล้ว ก็ไม่ต้องรันส่วนนี้ <br>
เปลี่ยนอีกและ 555

raise Exception("Skip this block")

for subset in ['Train','Test']:
    for p in provinces:
        data[subset][p].fillna(value=0.0, inplace=True)
        path = f"./data/{subset}/imputed_byzero/{p}_0imputed.csv"
        if not os.path.exists(path):
            data[subset][p].to_csv(path)
        else: print(path)

# ===========================================================

# Rename PM2.5 in Test set
for p in provinces:
    data['Test'][p].rename(columns={"PM2.5(µg/m3)":"PM2.5"}, inplace=True)

# Datetime that PM2.5 is missing
data_week_imputed = {"Train":{}}
for subset in ['Train','Test']:   # Since Test has no null-values
    for p in provinces:
        if subset == 'Test':
            data[subset][p] = data[subset][p].ffill().bfill()
            continue
        
        a = get_date_pm_missing(data[subset][p])
        day_imputed = impute_Avg_by(data[subset][p], 'day', a)

        a = get_date_pm_missing(day_imputed)
        week_imputed = impute_Avg_by(day_imputed, 'week', a)

        data_week_imputed[subset][p] = week_imputed

        print(f"{subset} : {p} : day_impute->week_impute -> remains {week_imputed['PM2.5'].isnull().sum()}")

# All provinces except Khon Kaen -> ffill
data_train_not_null = {}
for p in provinces:
    if p == 'Khon Kaen' : 
        # Deal with Khon Kaen separately later
        data_train_not_null[p] = data["Train"][p]
        continue
    data_train_not_null[p] = data_week_imputed["Train"][p].ffill().bfill()

# Assign back to data
data['Train'] = data_train_not_null

# After investigating Khon Kaen data, we'll keep only data after 2017, Nov.
khon_kaen = data['Train']['Khon Kaen'].loc["2017, 11":]

a = get_date_pm_missing(khon_kaen)
day_imputed = impute_Avg_by(khon_kaen, 'day', a)
print("Missing after day imputed : ", day_imputed.isnull().sum()[0])

a = get_date_pm_missing(day_imputed)
week_imputed = impute_Avg_by(day_imputed, 'week', a)
print("Missing after week imputed : ", week_imputed.isnull().sum()[0])

data_week_imputed["Train"]["Khon Kaen"] = week_imputed

data['Train']['Khon Kaen'] = data_week_imputed["Train"]["Khon Kaen"]

# Last re-check
for subset in ['Train','Test']:
    print(subset)
    for p in provinces:
        print(p)
        print(f"{data[subset][p].isnull().sum()}")

# Ok forget to ffill the Khon Kaen
data['Train']['Khon Kaen'].ffill(inplace=True)

for subset in ['Train', 'Test']:
    for p in provinces:
        path = f"./data/{subset}/imputed/{p}_imputed.csv"
        if not os.path.exists(path):
            data[subset][p].to_csv(path)
        else: print(path)

# xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx

raise Exception("Don't run below, it's deprecated. You can run next notebook")

def check_missing(data):
    missing = dict()
    for province in provinces:
        null_idx = data[province]['PM2.5'].isnull()
        missing[province] = data[province].loc[null_idx].index

        print(f'{province} has {len(missing[province])} missing values')
    return missing

missing = check_missing(data)

def impute_DayAvg(data, method):
    for province in provinces:

        if method == 'day':
            avg = data[province]['PM2.5'].resample('D').mean()
            
            for e in missing[province]:
                day = e.strftime("%Y-%m-%d")
                data[province].at[e,'PM2.5'] = avg.loc[day]

        elif method == 'week':
            avg = data[province]['PM2.5'].resample('W').mean()
            avg.index = avg.index.strftime('%Y-%U')

            for e in missing[province]:
                yr_week = e.strftime("%Y-%U")
                data[province].at[e,'PM2.5'] = avg.loc[yr_week]
    return data

## Fill with Avg. in that day

data = impute_DayAvg(data, method='day')
missing = check_missing(data)

## Fill with Avg. in that week

data = impute_DayAvg(data, method='week')
missing = check_missing(data)

## Forward fill

for province in provinces:
    if province == "Khon Kaen": continue
    data[province]['PM2.5'].fillna(method='ffill', inplace=True)

missing = check_missing(data)

data['Bangkok'].tail()

---

## save the imputed files

print(f'subset : {subset}')
for province in provinces:
    if province == "Khon Kaen": continue
    data[province].fillna(method='ffill').to_csv(f'.\data\{subset}\{province}_imputed.csv')


# Null value in others

for province in provinces:
    print(f"{province}\n", data[province].isnull().sum())
