==> 4. Fire.ipynb <==
from collections import defaultdict
import pandas as pd
import numpy as np
import os
from datetime import timedelta
import json

# Extract Fire data
There are 2 types of fire data
- Country labeled dataset
- Unlabeled dataset

This first part will deal with labeled one.

years = ['2016','2017','2018','2019']
countries = ['Cambodia','Myanmar','Thailand','Lao_PDR']
provinces = ['Bangkok','Chanthaburi','Chiang Mai','Kanchanaburi','Khon Kaen','Songkhla']

fire = {'2016':{}, '2017':{}, '2018':{}, '2019':{}}

left_join = dict(left_index=True, right_index=True, how='left')

Load all the data (all years, all countries) and transform them.

fire_2016_2019 = {}

for year in years:
    # Create full timeline
    timeline = pd.date_range(start=f'{year}-01-01', end=f'{year}-12-31 23:00:00', freq='H')
    left = pd.DataFrame(index=timeline)
    for country in countries:

        # Import data & manually parse dates
        df = pd.read_csv(f'../Fire hotspot/viirs-snpp_{year}_{country}.csv')
        df['hour'] = df['acq_time'].apply(str).str[:-2]
        df['datetime'] = pd.to_datetime(df['acq_date']+'T'+df['hour'].str.zfill(2), 
                            format='%Y-%m-%dT%H')   

        # Assume : frp represent brightness
        df = df[['latitude','longitude','frp','datetime']]
        df.rename(columns={'frp':country+'_frp'}, inplace=True)
        df.rename(columns={'latitude':country+'_lat'}, inplace=True)
        df.rename(columns={'longitude':country+'_long'}, inplace=True)

        # Assume : ไม่มีค่าไฟใน record = ไม่มีไฟ -> frp = 0
        df = df.groupby('datetime').mean()
        df = df.resample('H').mean().fillna(value=0.0)
        left = left.merge(df, **left_join)
    
    fire_2016_2019[year] = left.fillna(value=0.0)

Assemble them as the <u><strong>fire data for 2016 - 2019</strong></u>. And record the max, min of latitude, longitude of each country for further use.

# 2016-2019 fire data
fire_2016_2019 = pd.concat(fire_2016_2019, axis=0)
fire_2016_2019.index = fire_2016_2019.index.levels[1]

# Find max, min of latitude, longitude of each country
records = {}
will_drop = []
for c in countries:
    temporary = fire_2016_2019[[c+'_lat', c+'_long']].replace(0.0, np.nan)
    records[c] = {'lat':
                    {'min':temporary[c+'_lat'].min(),
                     'max':temporary[c+'_lat'].max()},
                  'long':
                    {'min':temporary[c+'_long'].min(),
                     'max':temporary[c+'_long'].max()}
                  }
    will_drop += [c+'_lat', c+'_long']

# Drop unused columns
fire_2016_2019.drop(will_drop, inplace=True, axis=1)

# 2016-2019 fire data
fire_2016_2019
# Once we have 2016-2019 data, we will find 2019-2020 data and append to it.

# We will use these min,max of lat,long to identify the country of unlabeled data
print(json.dumps(records, indent=3, sort_keys=True))

fire_files = [ 'fire_archive_V1_163551.csv', 'fire_nrt_J1V-C2_163550.csv']

# Read two 2020 fire files
fire_2020 = {}
for f in fire_files:
    df = pd.read_csv(f'../Fire hotspot/{f}')
    df['hour'] = df['acq_time'].apply(str).str[:-2]
    df['datetime'] = pd.to_datetime(
                                df['acq_date']+'T'+df['hour'].str.zfill(2), 
                                format='%Y-%m-%dT%H')
    
    df = df.groupby(by='datetime').mean()
    fire_2020[f] = df

import geopy
from geopy.geocoders import Nominatim

def assign_country(x):
    lati, longi = x[0], x[1]

    # If lat-long is clear
    choice = []
    for c in countries:
        lat_const = lati<records[c]['lat']['max'] and lati>records[c]['lat']['min']
        long_const = longi<records[c]['long']['max'] and longi>records[c]['long']['min']
        if lat_const and long_const:
            choice.append(c)
    if len(choice)==1:
        return choice[0]
    
    # If lat-long is not clear
    del choice
    geolocator = Nominatim(user_agent="app")
    try:
        location = geolocator.reverse([lati, longi])
        return location.raw['address']['country_code']
    except:
        return np.nan

# Identify country
if not os.path.exists('archive_V1_country_code.csv'):
    # Find country code 
    country_code = data[fire_files[0]].apply(lambda x: assign_country(x), axis=1)
    country_code.to_csv('archive_V1_country_code.csv')
else: print('archive_V1_country_code.csv is already')

if not os.path.exists('nrt_J1V-C2_country_code.csv'):
    # Find country code 
    country_code = data[fire_files[1]].apply(lambda x: assign_country(x), axis=1)
    country_code.to_csv('nrt_J1V-C2_country_code.csv')
else: print('nrt_J1V-C2_country_code.csv is already')

code = {'Thailand':'th', 'Myanmar':'mm', 'Lao_PDR':'la', 'Cambodia':'kh'}
focus = {}
for f in fire_files:
    # Read country code file
    if f.split('_')[1] == 'archive':
        country_code_map = pd.read_csv('archive_V1_country_code.csv', parse_dates=True, index_col=0)
    elif f.split('_')[1] == 'nrt':
        country_code_map = pd.read_csv('nrt_J1V-C2_country_code.csv', parse_dates=True, index_col=0)
    
    country_code_map.columns = ['country_code']
    country_code_map.replace(code, inplace=True)

    # Filter out out-of-interest countries
    mask = country_code_map['country_code'].isin(code.values())
    focus[f] = country_code_map[mask]

# use fire_files[0]
file_0 = focus[fire_files[0]].loc["2020":"2020-03-18 20:00:00"]
file_1 = focus[fire_files[1]].loc["2020":"2020-03-18 20:00:00"]
assert len(file_0)>len(file_1)

fire_2020_used = fire_2020[fire_files[0]].loc[file_0.index]
fire_2020_used = fire_2020_used.merge(file_0, **left_join)
fire_2020_used = fire_2020_used[['frp', 'country_code']]
fire_2020_used

# Pivoting to right format
fire_2020 = pd.pivot(fire_2020_used.reset_index(), 
                index='datetime', columns = 'country_code', values='frp')

fire_2020.rename(columns={'kh':'Cambodia_frp', 
                        'la':'Lao_PDR_frp', 
                        'mm':'Myanmar_frp', 
                        'th':'Thailand_frp'}, inplace=True)
fire_2020.fillna(value=0.0, inplace=True)
fire_2020.tail()

# Create full time line
timeline_2020 = pd.DataFrame(index=pd.date_range('2020-01-01 00:00:00', '2020-03-18 18:00:00', freq='H'))
fire_2020 = timeline_2020.merge(fire_2020, **left_join).fillna(value=0.0)

# Append to 2016-2019 fire
fire_all = fire_2016_2019.append(fire_2020)

# Localize 
fire_all.index = fire_all.index + timedelta(hours=7)
fire_all_localized = fire_all.copy()

if not os.path.exists("./data/fire_all_localized.csv"):
    fire_all_localized.to_csv("./data/fire_all_localized.csv")
else: print("fire_all_localized is ready")

fire_all_localized

## Add fire to Train, Test of all provinces

fire_all_localized = pd.read_csv("./data/fire_all_localized.csv", parse_dates=True, index_col=0)

full_data = defaultdict(lambda: {})

for subset in ['Train','Test']:
    for p in provinces:
        df = pd.read_csv(f"./data/{subset}/imputed/{p}_imputed.csv", 
                                    index_col=0, parse_dates=True)
        if subset=='Test':
            df.rename(columns={"PM2.5(µg/m3)":"PM2.5"}, inplace=True)
        full_data[subset][p] = df.merge(fire_all_localized, **left_join)

for subset in ['Train','Test']:
    for p in provinces:
        path = f"./data/{subset}/imputed_fired/{p}_imputed_fired.csv"
        if not os.path.exists(path):
            full_data[subset][p].to_csv(path)
        else: print(path)

for p in provinces:
    print(full_data["Train"][p].shape)

# xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx

raise Exception("Don't run below, it's deprecated. You can run next notebook")

Next, we'll localize fire data to Thailand local timezone (UTC+7). Then, we integrate fire data to full dataset which we cleaned in `3.fillna.ipynb`.

provinces = ['Bangkok','Chanthaburi','Chiang Mai','Kanchanaburi','Khon Kaen','Songkhla']

# +7 Hours to localize the time
fire_all.index = fire_all.index + timedelta(hours=7)

# Integrate to full,imputed data
mega = {}
for province in provinces:
    if province == 'Khon Kaen' : continue
    df = pd.read_csv(f"./data/Train/imputed/{province}_imputed.csv", parse_dates=True, index_col=0)
    mega[province] = df.merge(fire_all, left_index=True, right_index=True, how='left')

mega

This is an example of *preprocessed* data.

# This is data integrated with localized fire
mega['Bangkok'].loc['2016-3-3']

Lastly, we <u>save</u> it the *preprocessed* dataset ready to be fed to the model.

# Save imputed, fire integrated data
for province in provinces:
    if province == 'Khon Kaen' : continue
    path = f'./data/Train/fire_integrated/{province}_fire_integrated.csv'
    if not os.path.exists(path):
        mega[province].to_csv(path)
    else:
        print(f"{province} already")

save ข้อมูลไฟทั้งหมด เก็บไว้ใช้ต่อกับ test set

import glob

# Save 2016-2019 fire data
if len(glob.glob("./data/fire_2016_to_2019_localize.csv"))==0:
    fire_all.to_csv('./data/fire_2016_to_2019_localize.csv')
else:
    print('already')

# Expand Fire data
In this project, the model will be evaluated using **test data** which is the data from `2019-03-18 08:00:00` to `2020-03-18 20:00:00`. So, we need to find more fire data since the labeled data we had ends at 2019-12-31.

We have several unlabeled fire data. So, We have to choose one. First we import 2016-2019 fire data.

localized_2016_2019 = pd.read_csv('./data/fire_2016_to_2019_localize.csv', index_col=0, parse_dates=True)
localized_2016_2019.tail()

## Lat-long info from viirs-snpp data
We recorded the latitude-longitude from the labeled dataset. We will use it to identify the country of unlabeled dataset.

records

After inspecting the dataset, we see that these 2 dataset `fire_archive_V1_163551.csv`,` fire_nrt_J1V-C2_163550.csv` have the information in date range that we're interested in.

fire_files = [ 'fire_archive_V1_163551.csv', 'fire_nrt_J1V-C2_163550.csv']

# Read two 2020 fire files
data = {}
for f in fire_files:
    df = pd.read_csv(f'./Fire hotspot/{f}')
    df['hour'] = df['acq_time'].apply(str).str[:-2]
    df['datetime'] = pd.to_datetime(
                                df['acq_date']+'T'+df['hour'].str.zfill(2), 
                                format='%Y-%m-%dT%H')
    
    df = df.groupby(by='datetime').mean()
    data[f] = df

## Ambiguity
Identifying the country from max,min of latitude,longitude may ambiguous since those values might overlaped. We address this issue by using external API to find the exact location of latitude, longitude. We'll use `geopy` package.

import geopy
from geopy.geocoders import Nominatim

def assign_country(x):
    lati, longi = x[0], x[1]

    # If lat-long is clear
    choice = []
    for c in countries:
        lat_const = lati<records[c]['lat']['max'] and lati>records[c]['lat']['min']
        long_const = longi<records[c]['long']['max'] and longi>records[c]['long']['min']
        if lat_const and long_const:
            choice.append(c)
    if len(choice)==1:
        return choice[0]
    
    # If lat-long is not clear
    del choice
    geolocator = Nominatim(user_agent="app")
    try:
        location = geolocator.reverse([lati, longi])
        return location.raw['address']['country_code']
    except:
        return np.nan

It took so long to find all country code from latitude, longitude. So, we decide to keep the code.

# Identify country
if not os.path.exists('archive_V1_country_code.csv'):
    # Find country code 
    country_code = data[fire_files[0]].apply(lambda x: assign_country(x), axis=1)
    country_code.to_csv('archive_V1_country_code.csv')
else: print('archive_V1_country_code.csv is already')

if not os.path.exists('nrt_J1V-C2_country_code.csv'):
    # Find country code 
    country_code = data[fire_files[1]].apply(lambda x: assign_country(x), axis=1)
    country_code.to_csv('nrt_J1V-C2_country_code.csv')
else: print('nrt_J1V-C2_country_code.csv is already')

Assign country code back to the `data[0], data[1]`

code = {'Thailand':'th', 'Myanmar':'mm', 'Lao_PDR':'la', 'Cambodia':'kh'}
focus = {}
for f in fire_files:
    # Read country code file
    if f.split('_')[1] == 'archive':
        country_code = pd.read_csv('archive_V1_country_code.csv', parse_dates=True, index_col=0)
    elif f.split('_')[1] == 'nrt':
        country_code = pd.read_csv('nrt_J1V-C2_country_code.csv', parse_dates=True, index_col=0)
    country_code.columns = ['country_code']

    # Assign back
    data[f]['country_code'] = country_code
    data[f].replace(code, inplace=True)

    # Filter out out-of-interest countries
    mask = data[f]['country_code'].isin(code.values())
    focus[f] = data[f][mask]

Next, we'll append 2020 fire data to localized 2016-2019 data. But first, we have to localize it.

# Pivoting to right format
selected = focus[fire_files[0]].loc['2020-01-01 00:00:00':,['country_code','frp']].reset_index()
selected = pd.pivot(selected, index='datetime', columns = 'country_code', values='frp')
selected.rename(columns={'kh':'Cambodia_frp', 
                        'la':'Lao_PDR_frp', 
                        'mm':'Myanmar_frp', 
                        'th':'Thailand_frp'}, inplace=True)
selected.head()

# Create full time line
extended_timeline = pd.DataFrame(index=pd.date_range('2020-01-01 00:00:00', '2020-05-31 06:00:00', freq='H'))
selected = extended_timeline.merge(selected, left_index=True, right_index=True, how='left').fillna(0.0)

# Localize 
selected.index = selected.index + timedelta(hours=7)

localized_2020 = selected.copy()
localized_2020

localized_2016_2019.tail()

localized_2016_2020 = localized_2016_2019.append(localized_2020)
localized_2016_2020

if not os.path.exists('./data/fire_2016_to_2020_localize.csv'):
    localized_2016_2020.to_csv('./data/fire_2016_to_2020_localize.csv')
else:
    print('Full fire already')

## Test: fire integrated
Once full fire is ready, we integrate it to the `Test set`.

full_fire_localized = pd.read_csv('./data/fire_2016_to_2020_localize.csv', parse_dates=True, index_col=0)
full_fire_localized

# Integrate to full,imputed Test data
mega = {}
for province in provinces:
    if province == 'Khon Kaen' : continue

    # Read imputed data & Integrate the localized fire
    df = pd.read_csv(f"./data/Test/{province}_imputed.csv", parse_dates=True, index_col=0)
    mega[province] = df.merge(full_fire_localized, left_index=True, right_index=True, how='left')

    # Save imputed, fire integrated Test data
    path = f'./data/Test/fire_integrated/{province}_fire_integrated.csv'
    if not os.path.exists(path):
        mega[province].to_csv(path)
    else:
        print(f"{province} already")

mega[province]

mega[province]
