==> 7. submission.ipynb <==
from util_input_output_model import *
from collections import defaultdict
from datetime import timedelta
import tensorflow as tf
import numpy as np
import pandas as pd
import time
provinces = ['Bangkok','Chanthaburi','Chiang Mai','Kanchanaburi','Songkhla','Khon Kaen']

features = ['PM2.5','WindDir','Wind Speed(km/h)','Temp(C)',
            'Cambodia_frp','Myanmar_frp','Thailand_frp','Lao_PDR_frp']
tf.__version__

# 1) Preparing data

timesteps = 720
print(f'timesteps = {timesteps}')

# Takes 5-10 min
feature_used = ['PM2.5','WindDir','Wind Speed(km/h)','Temp(C)']
Train_data, X_train, Y_train = prepare_train_data(timesteps, feature_used = feature_used)
Test_data, X_test, Y_test = prepare_new_test(Train_data, timesteps, feature_used = feature_used)

# 2) Standardize the data


# Scale train set
x_train_scalers, y_train_scalers, X_train_scaled, Y_train_scaled = scale_data(X_train, Y_train,Train_data)

# Scale test set
X_test_scaled = defaultdict(lambda: [])
Y_test_scaled = defaultdict(lambda: [])

for p in provinces:
    print(p)
    for e in X_test[p]:
        X_test_scaled[p].append(x_train_scalers[p].transform(e))
    for e in Y_test[p]:
        Y_test_scaled[p].append(y_train_scalers[p].transform(e))

# 3) Reshaping

x_, y_ = {}, {}

for p in provinces:
    x_[p] = {"Train": np.array(X_train_scaled[p]),
             "Test": np.array(X_test_scaled[p])}
    
    y_[p] = {"Train": np.array(Y_train_scaled[p]).squeeze(axis=2),
             "Test": np.array(Y_test_scaled[p]).squeeze(axis=2)}

    print(p)
    print(f'X Train: {x_[p]["Train"].shape}')
    print(f'Y Train: {y_[p]["Train"].shape}')
    print(f'X Test: {x_[p]["Test"].shape}')
    print(f'Y Test: {y_[p]["Test"].shape}\n')

# 4) Load the model

These models below have *720* timesteps as an input, each timestep contains 4 features(PM2.5, Wind direction, Wind Speed, Temp) and give RMSE = **10.56**

    - Bangkok : Models\Bangkok_run_2021_05_07-20_31_15_chckpoint.h5
    - Chanthaburi : Models\Chanthaburi_run_2021_05_08-12_58_02_chckpoint.h5
    - Songkhla : Models\Songkhla_run_2021_05_08-19_56_09_chckpoint.h5
    - Kanchanaburi : Models\Kanchanaburi_run_2021_05_08-15_58_45_chckpoint.h5
    - Khon Kaen : Models\Khon Kaen_run_2021_05_07-23_38_36_chckpoint.h5
    - Chiang Mai : Models\Chiang Mai_run_2021_05_08-07_41_04_chckpoint.h5
        
These models below have *360* timesteps as an input, each timestep contains all features(PM2.5, Wind direction, Wind Speed, Temp, fire from 4 countries) and give RMSE = **10.96**

    - Bangkok : Models\Bangkok_run_2021_04_30-14_43_39.h5
    - Chanthaburi : Models\Chanthaburi_run_2021_04_26-10_42_58_final_26Apr1155.h5
    - Songkhla : Models\Songkhla_run_2021_04_25-22_41_22.h5
    - Kanchanaburi : Models\Kanchanaburi_run_2021_04_30-16_22_11.h5
    - Khon Kaen : Models\Khon Kaen_run_2021_04_27-16_20_46_kind_of_final.h5
    - Chiang Mai : Models\Chiang Mai_run_2021_04_30-15_37_01.h5

# 5) Submission

models = {'Bangkok':'Models\Bangkok_run_2021_05_07-20_31_15_chckpoint.h5',
'Chanthaburi':'Models\Chanthaburi_run_2021_05_08-12_58_02_chckpoint.h5',
'Songkhla':'Models\Songkhla_run_2021_05_08-19_56_09_chckpoint.h5',
'Kanchanaburi':'Models\Kanchanaburi_run_2021_05_08-15_58_45_chckpoint.h5',
'Khon Kaen':'Models\Khon Kaen_run_2021_05_07-23_38_36_chckpoint.h5',
'Chiang Mai':'Models\Chiang Mai_run_2021_05_08-07_41_04_chckpoint.h5'}
grand = []
for province in ["Chanthaburi","Chiang Mai","Kanchanaburi","Bangkok","Khon Kaen","Songkhla"]:

    # Load the trained model
    model = tf.keras.models.load_model(models[province])

    # Load Test data
    x_eval = x_[province]['Test']

    # Predict Test data
    pred = model(x_eval)

    # Store predictions in array
    prediction_1D = []
    for e in pred:
        prediction_1D.append(y_train_scalers[province].inverse_transform(e.numpy().reshape((-1,1))))

    prediction_1D = np.array(prediction_1D).reshape((-1,1))
    province_prediction = pd.DataFrame({'Predicted':prediction_1D.ravel()})

    grand.append(province_prediction)
    print(f"{province} done")

# This is what the prediction look like
pd.concat(grand, ignore_index=True)

# Save the result
submit = pd.concat(grand, ignore_index=True).rename_axis('Id') #.to_csv("./predictions/sub3.csv")

## 5.1) make_submission
ในส่วนนี้คัดลอกมาจากไฟล์ `make_submission.ipynb` ที่พี่ TA ให้ใหม่หลังจากแก้ submission ใน Kaggle

miss_index = pd.read_csv('./predictions/missing_index.csv')
prediction_index = pd.read_csv('./predictions/output_index.csv')

submit['datetime index'] = prediction_index['predicted timestamp']
submit[['Predicted','datetime index']]

cond = submit['datetime index'].isin(miss_index['Missing date'])
submit.drop(submit[cond].index, inplace = True)

submit.reset_index(inplace=True)
submit.drop(columns={'Id'},inplace=True)
submit

submit['id'] = submit.index
submit = submit[['Predicted','id']]

submit.set_index('id',inplace=True)
submit

import time
time_now = time.strftime("%Y_%b_%d-%H_%M_%S")
submit.to_csv(f'./predictions/submission_{time_now}.csv')
